{"schema":"decodex/commit/1","summary":"Split docs extension acceptance modules","authority":"manual"}
